requirements.txt
streamlit
groq
python-dotenv
pdf2image